requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pydantic
python-dotenv
qrcode
pillow
numpy
opencv-python
onnxruntime
insightface
greenlet